feat(frontend): adds `/login` page (#45)
* feat(frontend-style): moved tailwind theme and custom utilities into dedicated files

* refactor(frontend): moved single form pages into one common layout `/(form)/..`

* feat(frontend-auth): add `/login` page
