+++ packages/frontend/biome.jsonc
@@ -5,6 +5,9 @@
     "clientKind": "git",
     "useIgnoreFile": true
   },
+  "css": {
+    "parser": { "tailwindDirectives": true }
+  },
   "files": {
     "ignoreUnknown": false
   },
