+++ packages/frontend/src/lib/styles/utilities.css
@@ -1,0 +1,10 @@
+@utility form {
+  @apply card;
+  @apply bg-surface-50-950!;
+  @apply border border-surface-200-800;
+  @apply ring-6 ring-surface-100-900;
+  @apply outline outline-offset-6 outline-surface-200-800;
+  @apply m-1.75;
+
+  @apply flex flex-col items-center gap-4 p-6 *:w-full;
+}
