+++ packages/frontend/src/routes/(form)/setup/bot/components/Deployment.svelte
