+++ packages/frontend/src/routes/(form)/setup/bot/components/Scaling.svelte
