+++ packages/frontend/svelte.config.js
@@ -14,6 +14,9 @@ const config = {
     adapter: adapter(),
     experimental: { remoteFunctions: true },
   },
+  compilerOptions: {
+    experimental: { async: true },
+  },
 };
 
 export default config;
